+++ Scripts/Editor/SoftMaskEditor.cs
@@ -14,9 +14,9 @@ namespace Coffee.UISoftMask
     [CanEditMultipleObjects]
     public class SoftMaskEditor : Editor
     {
-        const string k_PrefsPreview = "SoftMaskEditor_Preview";
-        static readonly List<Graphic> s_Graphics = new List<Graphic>();
-        static bool s_Preview;
+        private const string k_PrefsPreview = "SoftMaskEditor_Preview";
+        private static readonly List<Graphic> s_Graphics = new List<Graphic>();
+        private static bool s_Preview;
 
         private void OnEnable()
         {
@@ -35,9 +35,7 @@ namespace Coffee.UISoftMask
             if (0 < fixTargets.Count)
             {
                 GUILayout.BeginHorizontal();
-                EditorGUILayout.HelpBox(
-                    "There are child Graphics that does not have a SoftMaskable component.\nAdd SoftMaskable component to them.",
-                    MessageType.Warning);
+                EditorGUILayout.HelpBox("There are child Graphics that does not have a SoftMaskable component.\nAdd SoftMaskable component to them.", MessageType.Warning);
                 GUILayout.BeginVertical();
                 if (GUILayout.Button("Fix"))
                 {
@@ -59,9 +57,8 @@ namespace Coffee.UISoftMask
             }
 
             // Preview buffer.
-            GUILayout.BeginHorizontal(EditorStyles.helpBox);
-            if (s_Preview != (s_Preview = EditorGUILayout.ToggleLeft("Preview Buffer", s_Preview,
-                GUILayout.MaxWidth(EditorGUIUtility.labelWidth))))
+            GUILayout.BeginVertical(EditorStyles.helpBox);
+            if (s_Preview != (s_Preview = EditorGUILayout.ToggleLeft("Preview Buffer", s_Preview)))
             {
                 EditorPrefs.SetBool(k_PrefsPreview, s_Preview);
             }
@@ -69,37 +66,35 @@ namespace Coffee.UISoftMask
             if (s_Preview)
             {
                 var tex = current.softMaskBuffer;
-                var width = tex.width * 64 / tex.height;
-                EditorGUI.DrawPreviewTexture(GUILayoutUtility.GetRect(width, 64), tex, null, ScaleMode.ScaleToFit);
+                var width = tex.width * 128 / tex.height;
+                EditorGUI.DrawPreviewTexture(GUILayoutUtility.GetRect(width, 128), tex, null, ScaleMode.ScaleToFit);
                 Repaint();
             }
-
-            GUILayout.FlexibleSpace();
-            GUILayout.EndHorizontal();
+            GUILayout.EndVertical();
         }
 
 
         //%%%% Context menu for editor %%%%
         [MenuItem("CONTEXT/Mask/Convert To SoftMask", true)]
-        static bool _ConvertToSoftMask(MenuCommand command)
+        private static bool _ConvertToSoftMask(MenuCommand command)
         {
             return EditorUtils.CanConvertTo<SoftMask>(command.context);
         }
 
         [MenuItem("CONTEXT/Mask/Convert To SoftMask", false)]
-        static void ConvertToSoftMask(MenuCommand command)
+        private static void ConvertToSoftMask(MenuCommand command)
         {
             EditorUtils.ConvertTo<SoftMask>(command.context);
         }
 
         [MenuItem("CONTEXT/Mask/Convert To Mask", true)]
-        static bool _ConvertToMask(MenuCommand command)
+        private static bool _ConvertToMask(MenuCommand command)
         {
             return EditorUtils.CanConvertTo<Mask>(command.context);
         }
 
         [MenuItem("CONTEXT/Mask/Convert To Mask", false)]
-        static void ConvertToMask(MenuCommand command)
+        private static void ConvertToMask(MenuCommand command)
         {
             EditorUtils.ConvertTo<Mask>(command.context);
         }